:art: keep functions close to source
